'下载相关修改文件'

+++ app/src/main/java/com/xiaoe/shop/wxb/base/XiaoeActivity.java
@@ -547,9 +547,13 @@ public class XiaoeActivity extends SwipeBackActivity implements INetworkResponse
     public void onClickConfirm(View view, int tag) {
         if(tag == DIALOG_TAG_LOADING){
             SQLiteUtil liteUtil = SQLiteUtil.init(this,  new LoginSQLiteCallback());
-            liteUtil.execSQL("delete from " + LoginSQLiteCallback.TABLE_NAME_USER);
+            if(liteUtil.tabIsExist(LoginSQLiteCallback.TABLE_NAME_USER)){
+                liteUtil.execSQL("delete from " + LoginSQLiteCallback.TABLE_NAME_USER);
+            }
             SQLiteUtil lrUtil = SQLiteUtil.init(this, new LrSQLiteCallback());
-            lrUtil.execSQL("delete from " + LrSQLiteCallback.TABLE_NAME_LR);
+            if (lrUtil.tabIsExist(LrSQLiteCallback.TABLE_NAME_LR)) {
+                lrUtil.execSQL("delete from " + LrSQLiteCallback.TABLE_NAME_LR);
+            }
             CommonUserInfo.getInstance().clearUserInfo();
             CommonUserInfo.getInstance().clearLoginUserInfo();
             CommonUserInfo.setApiToken("");
